requirements.txt
Pillow>=10.1.0
customtkinter>=5.2.0
numpy>=1.24.0
zstandard>=0.21.0
tkinterdnd2>=0.3.0
